requirements.txt
discord.py>=2.3.0
aiohttp>=3.8.0
aiosqlite>=0.19.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
APScheduler>=3.10.0